Add a "fill" option to the "fit" property
This allows a user to opt into their image being stretched, which FT
labs needs for a project they're working on.

I opted for `fill` instead of `stretch` because the documentation
explicitly references the
[CSS3 object-fit property](https://www.w3.org/TR/css3-images/#the-object-fit).
In order to keep following the same naming conventions, we need to
go with `fill`.

See #340 for more information.