store images in Cloudinary using content hash as their ID
This changes the Origami Image Service's use of Cloudinary's fetch API with their upload API and changes how the images are stored.

Previously we used the image's url as the public ID when storing on Cloudinary. This is now causing issues because the publishing platforms which use the image service have begun to use new image urls for the same image, meaning we have thousands upoin thousands of duplicate images being stored in Cloudinary.

This commit makes the Origami Image Service download the image first and calculate the SHA256 of the image, using the SHA value as the public ID. This solves the issue of duplicate images because the exact same image served at different URLs will now resolve to the same public ID in Cloudinary.

To set public ID's in cloudinary is not possible using the fetch API which is why this commit also switches to the upload API.