feat: replace fthead implementation which uses the deprecated headshot imageset with api calls to CAPI
The author headshots were moved from being owned by Origami to being owned by Editorial a couple years ago.

When looking at moving the image-sets into the image-service directly, we thought we could remove fthead and headshot-images completely.

This turned out to not be true, many products use fthead still as it is a very convienant way to retrieve an image for an author.

Instead of removing this capability from the image-service, we've replaced the implementation with one which uses the Concept and People APIs to find the up-to-date headshot image url for an author.

This approach means we can remove the headshot-images dependency and still have the convenient fthead custom scheme which many products are using.

This also unblocks the work to inline image-set into the image-service, as all the remaining image-sets are small enough to live within ther Heroku App directly