Fix svg tint integration test.
At some point "http://im.ft-static.com/content/images/[image]"
began to return a 301 to a https endpoint, with the wrong path
"https://im.ft-static.com/im-ft-static-com/content/images/[image]"
which returned a 403. This broke our integration test which relied
on a 200 response – to test a non-SVG image behaves as expected.

This updates the request to `https` rather than `http` to side-step
the issue.