docs: introduce container diagram
A rough first attempt at a container diagram based on the discussions in our meetings today. This definitely has more detail than is needed for the first level. But it's good that we got it all written down somewhere. The next step would be to use this information and split into the lower level diagrams.