fix: scripts/delete-old-images-from-cloudinary.js
To reduce Cloudinary costs we want to delete old images. We have a script to do this, which is run every day via a "Heroku scheduler" on origami-image-service-eu. However this script errors due to api limits, it's stuck trying to delete the same images on loop.

- Pass `next_cursor` to Cloudinary search, so the next results are returned correctly.
- Set `keep_original` when deleting images. This will delete only derived images – transforms – not the originals. We don't promise to keep originals and expect the source images to remain available, but I don't want to have to explain that to any angry product teams right now.
- Update logs to help us monitor the script. If it deletes many images and then fails before the script exits, we want to know.

We probably don't want to run this updated script every day as it will have to process every image we have stored, that is older than a month, everyday. This is because it no longer deletes the originals. Let's run this inititally to clear the backlog of old transforms, monitor the duration of future runs, and consider changing the search expression to process a limited range of images such as those uploaded between the last 3-6 months `.expression('uploaded_at:[24w TO 12w]')`.