chore: Remove redis hostname monitoring test.
This fails regularly as Heroku rotates credentials and we have not
automated updating these for Github Actions. We plan to remove the
feature under test and do not mind if it is broken, provided all
other tests pass. This is because it has done its job. We know
there are over 2,000 unique hosts for images and... we need to take
a different approach to limiting that than we originally thought.